refactoring and introduction of rest endpoint

+++ src/main/resources/Doctors.sql
@@ -1,11 +1,19 @@
-drop table Doctor;
+drop table Doctor if exists;
 create table Doctor(
-  ID int not null,
+  ID varchar(36) default(uuid()),
+  license int,
   NAME varchar(100) not null,
-  PRIMARY KEY ( ID )
+  PRIMARY KEY (ID)
 );
 
-INSERT INTO Doctor (id, name) VALUES (1245, 'Doctor Fisch');
-INSERT INTO Doctor (id, name) VALUES (66645, 'Doctor Hummel');
-INSERT INTO Doctor (id, name) VALUES (199985, 'Doctor Winter');
+drop index Doctor if exists;
+CREATE INDEX idx_doctor ON Doctor (ID);
+
+INSERT INTO Doctor (license, name) VALUES (1245, 'Doctor Fisch');
+INSERT INTO Doctor (license, name) VALUES (66645, 'Doctor Hummel');
+INSERT INTO Doctor (license, name) VALUES (199985, 'Doctor Winter');
+INSERT INTO Doctor (license, name) VALUES (1245, 'Doctor Fischer');
+INSERT INTO Doctor (license, name) VALUES (66645, 'Doctor Berlin');
+INSERT INTO Doctor (license, name) VALUES (199985, 'Doctor Sommer');
+INSERT INTO Doctor (license, name) VALUES (199985, 'Doctor Filler');
